Moved models to individual py files.
