+++ templates/package/package_detail.html
@@ -97,7 +97,7 @@
             </tr>
             {% else %}
             <tr>
-                <td colspan="5">No flights are available.</td>
+                <td colspan="7">No flights are available.</td>
             </tr>
             {% endif %}
