+++ templates/v2/mission/mission.html
@@ -262,7 +262,7 @@
 										<p class="media-body pb-3 pl-2 mb-0  lh-125 border-bottom border-gray">
 											<strong class="d-block text-gray-dark"><a href="{% url 'user_profile' comment.user.id %}" class="text-decoration-none text-reset">{{ comment.user.username}}</a></strong>
 											
-											{{ comment.comment }}
+											{{ comment.comment|urlize }}
 											<p class="small">{{ comment.date_created | timesince}} ago.</p>
 										</p>
 								</div>
